<feat> : get_path_args func, ft_exit fix
