+++ .vscode/launch.json
@@ -11,7 +11,8 @@
             "cwd": "${fileDirname}",
             "environment": [],
             "externalConsole": false,
-            "MIMode": "lldb"
+            "MIMode": "lldb",
+            "preLaunchTask": "debug"
         },
     ]
 }
